manager_service/src/requirements.txt
Flask==3.0.3
requests==2.32.3
minio==7.2.7
cassandra-driver==3.29.1
uuid
